move modal components to correct level
