+++ src/components/modal/modal.jsx
@@ -1,4 +1,4 @@
-import React, { useRef } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import styles from './modal.module.css';
 import {CloseIcon}  from '@ya.praktikum/react-developer-burger-ui-components';
